# config/api.yaml
openapi: 3.0.1
info:
  title: Web API Vật Tư
  description: N19DCCN112
  contact:
    email: mynth30@gmail.com
  version: 1
  license:
    name: MIT Licence
    url: https://github.com/thombergs/code-examples/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/Rates/{id}":
    get:
      tags:
        - Rates
      operationId: get1_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            "$ref": "#/components/schemas/RateId"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RateDTO"
    put:
      tags:
        - Rates
      operationId: update_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            "$ref": "#/components/schemas/RateId"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RateDTO"
        required: true
      responses:
        '202':
          description: Accepted
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RateDTO"
    delete:
      tags:
        - Rates
      operationId: delete_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            "$ref": "#/components/schemas/RateId"
      responses:
        '202':
          description: Accepted
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RateDTO"
...
  "/api/Rates":
    get:
      tags:
        - Rates
      operationId: getAll_5
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/RateDTO"
    post:
      tags:
        - Rates
      operationId: insert_5
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RateDTO"
        required: true
      responses:
        '201':
          description: Created
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RateDTO"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: integer
          format: int64
        description:
          type: string
        products:
          type: array
          items:
            "$ref": "#/components/schemas/Product"
    Event:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
        eventName:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        discountCode:
          type: string
        discountValue:
          type: integer
          format: int32
        eventStatus:
          type: string
          enum:
            - ACTICATED
            - PENDING_ACTICATION
            - EXPIRED
    EventProduct:
      type: object
      properties:
        eventProductId:
          "$ref": "#/components/schemas/EventProductId"
        createDate:
          type: string
          format: date-time
        updateDate:
          type: string
          format: date-time
    EventProductId:
      type: object
      properties:
        event:
          "$ref": "#/components/schemas/Event"
        products:
          "$ref": "#/components/schemas/Product"
...
    Product:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        createDate:
          type: string
          format: date-time
        description:
          type: string
        name:
          type: string
        price:
          type: integer
          format: int64
        remain:
          type: integer
          format: int64
        updateDate:
          type: string
          format: date-time
        expirationDate:
          type: string
          format: date-time
        category:
          "$ref": "#/components/schemas/Category"
        eventProducts:
          type: array
          items:
            "$ref": "#/components/schemas/EventProduct"
    Role:
      type: object
      properties:
        roleId:
          type: integer
          format: int64
        name:
          type: string
          enum:
            - ROLE_USER
            - ROLE_SHOP
        users:
          type: array
          items:
            "$ref": "#/components/schemas/User"
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
        username:
          type: string
        phone:
          type: string
        role:
          "$ref": "#/components/schemas/Role"
        userDetails:
          type: array
          items:
            "$ref": "#/components/schemas/UserDetail"
    UserDetail:
      type: object
      properties:
        userDetailId:
          type: integer
          format: int64
        address:
          type: string
        name:
          type: string
        addressDefault:
          type: integer
          format: int32
        userDetail:
          "$ref": "#/components/schemas/UserDetail"
...
    RateId:
      type: object
      properties:
        products:
          "$ref": "#/components/schemas/Product"
        user:
          "$ref": "#/components/schemas/User"
    RateDTO:
      required:
        - comment
      type: object
      properties:
        rateId:
          "$ref": "#/components/schemas/RateId"
        comment:
          type: string
        point:
          type: integer
          format: int32
...
  securitySchemes:
    brearerToken:
      type: http
      scheme: brearer
      bearerFormat: JWT